refactor(plugin): extract input_mailboxes.rs (#1079)
Moves the InputMailboxes callbacks + HOST_INPUT_MAILBOXES_VTABLE +
host_input_mailboxes_vtable() accessor + tier-1 test module out of
host_services/mod.rs into a sibling submodule. Removes the duplicate
write_extern_err helper from mod.rs (callers now reach it via
shared::wire::write_extern_err). mod.rs re-exports
HOST_INPUT_MAILBOXES_VTABLE (visibility widened from file-private
static to pub static) and host_input_mailboxes_vtable.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
