+++ libs/streamlib-engine/src/core/plugin/host_services/mod.rs
@@ -80,6 +80,7 @@ mod shared;
 
 mod audio_clock;
 mod gpu_context;
+mod input_mailboxes;
 mod output_writer;
 mod runtime_context;
 mod runtime_ops;
@@ -89,6 +90,7 @@ pub use gpu_context::{
     host_gpu_context_full_access_vtable, host_gpu_context_limited_access_vtable,
     HOST_GPU_CONTEXT_FULL_ACCESS_VTABLE, HOST_GPU_CONTEXT_LIMITED_ACCESS_VTABLE,
 };
+pub use input_mailboxes::{host_input_mailboxes_vtable, HOST_INPUT_MAILBOXES_VTABLE};
 pub use output_writer::{host_output_writer_vtable, HOST_OUTPUT_WRITER_VTABLE};
 pub use runtime_context::{host_runtime_context_vtable, HOST_RUNTIME_CONTEXT_VTABLE};
 pub use runtime_ops::{
@@ -1107,246 +1109,6 @@ fn emit_via_host_dispatch(
 // the audio-clock vtable is a `*const SharedAudioClock`, and `handle`
 // for the runtime-ops vtable is a `*const Arc<dyn RuntimeOperations>`.
 // The cdylib treats them all as opaque, dispatching through fn
-
-// =============================================================================
-// InputMailboxesVTable wrappers (issue #894)
-// =============================================================================
-
-/// SAFETY: caller must hand a `handle` that came from
-/// `Arc::into_raw(Arc<crate::iceoryx2::InputMailboxesInner>)`. The
-/// leaked strong count keeps the inner alive for the call's
-/// duration.
-unsafe fn handle_as_input_mailboxes_inner(
-    handle: *const c_void,
-) -> Option<&'static crate::iceoryx2::InputMailboxesInner> {
-    if handle.is_null() {
-        return None;
-    }
-    Some(unsafe { &*(handle as *const crate::iceoryx2::InputMailboxesInner) })
-}
-
-unsafe extern "C" fn host_input_mailboxes_read_raw(
-    handle: *const c_void,
-    port_ptr: *const u8,
-    port_len: usize,
-    out_buf: *mut u8,
-    out_cap: usize,
-    out_len: *mut usize,
-    out_timestamp: *mut i64,
-    has_data: *mut bool,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_input_mailboxes_read_raw",
-        || -> i32 {
-            if !out_len.is_null() {
-                unsafe {
-                    *out_len = 0;
-                }
-            }
-            if !has_data.is_null() {
-                unsafe {
-                    *has_data = false;
-                }
-            }
-            let Some(inner) = (unsafe { handle_as_input_mailboxes_inner(handle) }) else {
-                write_extern_err(
-                    "read_raw: null InputMailboxes handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if port_ptr.is_null() {
-                write_extern_err(
-                    "read_raw: null port_ptr",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let port_bytes = unsafe { std::slice::from_raw_parts(port_ptr, port_len) };
-            let port = match std::str::from_utf8(port_bytes) {
-                Ok(s) => s,
-                Err(e) => {
-                    write_extern_err(
-                        &format!("read_raw: port not UTF-8: {e}"),
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            // The inner's read_raw consumes the frame from the
-            // per-port mailbox. We can't peek non-destructively,
-            // so when the consumer's `out_buf` is too small we
-            // need to either (a) refuse and signal the required
-            // size or (b) consume + lose the frame. Choice (a) is
-            // safer; today's iceoryx2 max payload is bounded so
-            // a 4 KiB initial buffer covers ~all real traffic.
-            //
-            // The pattern: pop, measure, if it fits copy + signal
-            // success; if not, push it back to the head of the
-            // mailbox and signal required size. Today's
-            // PortMailbox doesn't expose push-to-head — the
-            // simplest sound implementation is to drain via
-            // `InputMailboxesInner::read_raw` (which already pops
-            // and returns the bytes), then if the bytes fit copy
-            // them; if not, return the required size and CONSUME
-            // the frame (the consumer's retry sees the next frame
-            // or empty mailbox). Truncation on overflow is the
-            // documented contract; in practice the β-shape's
-            // `read_raw` caller starts with a 4 KiB buffer and
-            // resizes proactively when the host indicates a
-            // larger payload, so the truncation path triggers
-            // only on the rare oversized inbound frame.
-            match inner.read_raw(port) {
-                Ok(Some((bytes, ts))) => {
-                    let required = bytes.len();
-                    if !has_data.is_null() {
-                        unsafe {
-                            *has_data = true;
-                        }
-                    }
-                    if !out_timestamp.is_null() {
-                        unsafe {
-                            *out_timestamp = ts;
-                        }
-                    }
-                    if !out_len.is_null() {
-                        unsafe {
-                            *out_len = required;
-                        }
-                    }
-                    if required <= out_cap && !out_buf.is_null() {
-                        unsafe {
-                            std::ptr::copy_nonoverlapping(bytes.as_ptr(), out_buf, required);
-                        }
-                    } else {
-                        // Truncation: indicate required size; the
-                        // consumer resizes and the next call sees
-                        // an empty mailbox (today's pop semantics
-                        // already consumed the frame). The
-                        // alternative (peek + non-destructive
-                        // size measurement) requires reworking
-                        // PortMailbox; deferred to a follow-up
-                        // when the truncation path is hit in
-                        // practice.
-                    }
-                    0
-                }
-                Ok(None) => 0, // has_data stays false
-                Err(e) => {
-                    write_extern_err(&e.to_string(), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_input_mailboxes_has_data(
-    handle: *const c_void,
-    port_ptr: *const u8,
-    port_len: usize,
-) -> bool {
-    run_host_extern_c(
-        "host_input_mailboxes_has_data",
-        || -> bool {
-            let Some(inner) = (unsafe { handle_as_input_mailboxes_inner(handle) }) else {
-                return false;
-            };
-            if port_ptr.is_null() {
-                return false;
-            }
-            let port_bytes = unsafe { std::slice::from_raw_parts(port_ptr, port_len) };
-            let Ok(port) = std::str::from_utf8(port_bytes) else {
-                return false;
-            };
-            inner.has_data(port)
-        },
-        false,
-    )
-}
-
-pub(crate) unsafe extern "C" fn host_input_mailboxes_clone_arc(
-    handle: *const c_void,
-) -> *const c_void {
-    run_host_extern_c(
-        "host_input_mailboxes_clone_arc",
-        || -> *const c_void {
-            if handle.is_null() {
-                return std::ptr::null();
-            }
-            // SAFETY: handle came from Arc::into_raw.
-            unsafe {
-                std::sync::Arc::<crate::iceoryx2::InputMailboxesInner>::increment_strong_count(
-                    handle as *const crate::iceoryx2::InputMailboxesInner,
-                );
-            }
-            handle
-        },
-        std::ptr::null(),
-    )
-}
-
-pub(crate) unsafe extern "C" fn host_input_mailboxes_drop_arc(handle: *const c_void) {
-    run_host_extern_c(
-        "host_input_mailboxes_drop_arc",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: handle came from Arc::into_raw.
-            unsafe {
-                std::sync::Arc::<crate::iceoryx2::InputMailboxesInner>::decrement_strong_count(
-                    handle as *const crate::iceoryx2::InputMailboxesInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-/// Per-DSO host-side static InputMailboxes dispatch table.
-static HOST_INPUT_MAILBOXES_VTABLE: streamlib_plugin_abi::InputMailboxesVTable =
-    streamlib_plugin_abi::InputMailboxesVTable {
-        layout_version: streamlib_plugin_abi::INPUT_MAILBOXES_VTABLE_LAYOUT_VERSION,
-        _reserved_padding: 0,
-        read_raw: host_input_mailboxes_read_raw,
-        has_data: host_input_mailboxes_has_data,
-        clone_arc: host_input_mailboxes_clone_arc,
-        drop_arc: host_input_mailboxes_drop_arc,
-    };
-
-/// Pointer to the [`streamlib_plugin_abi::InputMailboxesVTable`] this
-/// DSO should dispatch through.
-pub fn host_input_mailboxes_vtable() -> *const streamlib_plugin_abi::InputMailboxesVTable {
-    match host_callbacks() {
-        Some(c) if !c.input_mailboxes_vtable.is_null() => c.input_mailboxes_vtable,
-        _ => &HOST_INPUT_MAILBOXES_VTABLE,
-    }
-}
-
-/// Shared extern-C scratch err-buf writer for the OutputWriter +
-/// InputMailboxes host wrappers.
-fn write_extern_err(msg: &str, err_buf: *mut u8, err_buf_cap: usize, err_len: *mut usize) {
-    if err_buf.is_null() || err_len.is_null() {
-        return;
-    }
-    let bytes = msg.as_bytes();
-    let n = bytes.len().min(err_buf_cap);
-    unsafe {
-        std::ptr::copy_nonoverlapping(bytes.as_ptr(), err_buf, n);
-        *err_len = n;
-    }
-}
 
 // =============================================================================
 // runtime_facing — host-side payload builder
@@ -8633,178 +8395,6 @@ mod make_borrow_cached_field_regression_tests {
 /// touching the (null) handle.
 
 /// Tier-1 wire-format tests for [`HOST_INPUT_MAILBOXES_VTABLE`] (issue
-/// #894).
-#[cfg(test)]
-mod input_mailboxes_vtable_tier1_wire_format_tests {
-    use super::*;
-
-    #[test]
-    fn layout_version_matches_constant() {
-        assert_eq!(
-            HOST_INPUT_MAILBOXES_VTABLE.layout_version,
-            streamlib_plugin_abi::INPUT_MAILBOXES_VTABLE_LAYOUT_VERSION,
-        );
-    }
-
-    #[test]
-    fn read_raw_returns_error_on_null_handle() {
-        let mut buf = [0u8; 64];
-        let mut out_len = 0usize;
-        let mut out_ts = 0i64;
-        let mut has_data = false;
-        let mut err_buf = [0u8; 256];
-        let mut err_len = 0usize;
-        let port = b"any_port";
-        let rc = unsafe {
-            (HOST_INPUT_MAILBOXES_VTABLE.read_raw)(
-                std::ptr::null(),
-                port.as_ptr(),
-                port.len(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut out_len as *mut usize,
-                &mut out_ts as *mut i64,
-                &mut has_data as *mut bool,
-                err_buf.as_mut_ptr(),
-                err_buf.len(),
-                &mut err_len as *mut usize,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = std::str::from_utf8(&err_buf[..err_len]).unwrap();
-        assert!(
-            msg.contains("null InputMailboxes handle"),
-            "unexpected err message: {msg}"
-        );
-        assert!(!has_data);
-        assert_eq!(out_len, 0);
-    }
-
-    #[test]
-    fn read_raw_returns_error_on_invalid_utf8_port() {
-        let inner = std::sync::Arc::new(crate::iceoryx2::InputMailboxesInner::new());
-        let handle =
-            std::sync::Arc::into_raw(inner) as *const std::ffi::c_void;
-        let mut buf = [0u8; 64];
-        let mut out_len = 0usize;
-        let mut out_ts = 0i64;
-        let mut has_data = false;
-        let mut err_buf = [0u8; 256];
-        let mut err_len = 0usize;
-        let bad_port = b"\xff\xfe";
-        let rc = unsafe {
-            (HOST_INPUT_MAILBOXES_VTABLE.read_raw)(
-                handle,
-                bad_port.as_ptr(),
-                bad_port.len(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut out_len as *mut usize,
-                &mut out_ts as *mut i64,
-                &mut has_data as *mut bool,
-                err_buf.as_mut_ptr(),
-                err_buf.len(),
-                &mut err_len as *mut usize,
-            )
-        };
-        assert_eq!(rc, 1);
-        let msg = std::str::from_utf8(&err_buf[..err_len]).unwrap();
-        assert!(
-            msg.contains("port not UTF-8"),
-            "unexpected err message: {msg}"
-        );
-        unsafe {
-            std::sync::Arc::<crate::iceoryx2::InputMailboxesInner>::decrement_strong_count(
-                handle as *const _,
-            );
-        }
-    }
-
-    #[test]
-    fn read_raw_returns_no_data_on_empty_mailbox() {
-        let inner = std::sync::Arc::new(crate::iceoryx2::InputMailboxesInner::new());
-        inner.add_port(
-            "p",
-            8,
-            crate::iceoryx2::ReadMode::ReadNextInOrder,
-        );
-        let handle =
-            std::sync::Arc::into_raw(inner) as *const std::ffi::c_void;
-        let mut buf = [0u8; 64];
-        let mut out_len = 0usize;
-        let mut out_ts = 0i64;
-        let mut has_data = true; // start true to verify the wrapper sets it false
-        let mut err_buf = [0u8; 256];
-        let mut err_len = 0usize;
-        let port = b"p";
-        let rc = unsafe {
-            (HOST_INPUT_MAILBOXES_VTABLE.read_raw)(
-                handle,
-                port.as_ptr(),
-                port.len(),
-                buf.as_mut_ptr(),
-                buf.len(),
-                &mut out_len as *mut usize,
-                &mut out_ts as *mut i64,
-                &mut has_data as *mut bool,
-                err_buf.as_mut_ptr(),
-                err_buf.len(),
-                &mut err_len as *mut usize,
-            )
-        };
-        assert_eq!(rc, 0);
-        assert!(!has_data);
-        assert_eq!(out_len, 0);
-        unsafe {
-            std::sync::Arc::<crate::iceoryx2::InputMailboxesInner>::decrement_strong_count(
-                handle as *const _,
-            );
-        }
-    }
-
-    #[test]
-    fn has_data_returns_false_on_null_handle() {
-        let port = b"any";
-        let result = unsafe {
-            (HOST_INPUT_MAILBOXES_VTABLE.has_data)(
-                std::ptr::null(),
-                port.as_ptr(),
-                port.len(),
-            )
-        };
-        assert!(!result);
-    }
-
-    #[test]
-    fn clone_arc_returns_null_on_null_handle() {
-        let result = unsafe { (HOST_INPUT_MAILBOXES_VTABLE.clone_arc)(std::ptr::null()) };
-        assert!(result.is_null());
-    }
-
-    #[test]
-    fn drop_arc_is_noop_on_null_handle() {
-        unsafe { (HOST_INPUT_MAILBOXES_VTABLE.drop_arc)(std::ptr::null()) };
-    }
-
-    /// End-to-end refcount accounting: clone_arc bumps strong count
-    /// and returns the same handle; drop_arc decrements. Mirrors the
-    /// OutputWriter sibling test.
-    #[test]
-    fn clone_drop_arc_balance_strong_count() {
-        let inner = std::sync::Arc::new(crate::iceoryx2::InputMailboxesInner::new());
-        let inner_for_test = inner.clone();
-        let raw =
-            std::sync::Arc::into_raw(inner) as *const std::ffi::c_void;
-        assert_eq!(std::sync::Arc::strong_count(&inner_for_test), 2);
-        let cloned = unsafe { (HOST_INPUT_MAILBOXES_VTABLE.clone_arc)(raw) };
-        assert_eq!(cloned, raw);
-        assert_eq!(std::sync::Arc::strong_count(&inner_for_test), 3);
-        unsafe { (HOST_INPUT_MAILBOXES_VTABLE.drop_arc)(cloned) };
-        assert_eq!(std::sync::Arc::strong_count(&inner_for_test), 2);
-        unsafe { (HOST_INPUT_MAILBOXES_VTABLE.drop_arc)(raw) };
-        assert_eq!(std::sync::Arc::strong_count(&inner_for_test), 1);
-    }
-}
 
 #[cfg(all(test, target_os = "linux"))]
 mod rhi_color_converter_methods_vtable_tier1_wire_format_tests {
